% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect-explore.R
\name{compute_relative_log_expression}
\alias{compute_relative_log_expression}
\title{Computing relative log expression}
\usage{
compute_relative_log_expression(
  x,
  top_n,
  log2_transform = TRUE,
  log2_offset = 1
)
}
\arguments{
\item{x}{A matrix.}

\item{top_n}{an integer of the most variable N loci for relative log
expression (RLE) analysis.}

\item{log2_transform}{perform log2 transform or not.}

\item{log2_offset}{offset by certain value to avoid zeros. Default to 1.}
}
\value{
a matrix.
}
\description{
Computing relative log expression
}
