% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{cor_mtest}
\alias{cor_mtest}
\title{Computing the p-value of correlations.}
\usage{
cor_mtest(x, ...)
}
\arguments{
\item{x}{A matrix.}

\item{...}{Any arguments to pass to \code{\link[stats]{cor.test}}
function.}
}
\value{
A matrix of p-values.
}
\description{
Computing the p-value of correlations.
}
\note{
I borrow most of the codes from \href{http://www.sthda.com/english/wiki/visualize-correlation-matrix-using-correlogram}{STHDA}.
}
