% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{correlogram}
\alias{correlogram}
\title{Correlogram.}
\usage{
correlogram(x, pheno, ...)
}
\arguments{
\item{x}{A matrix.}

\item{...}{Any arguments passed to \code{\link[corrplot]{corrplot}}.}
}
\value{
(Invisibly) returns a reordered correlation matrix.
}
\description{
Correlogram.
}
