% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{diagnose_tsne.cor}
\alias{diagnose_tsne.cor}
\title{Diagnose multiple tSNE runs with the correlations of coordinates.}
\usage{
diagnose_tsne.cor(x, pheno, ...)
}
\arguments{
\item{x}{A list of the returned value of \code{\link[RTsne]{RTsne}}
which itself is a list.}

\item{pheno}{A character or factor vector.}

\item{...}{Any arguments passed to \code{\link[ggplot2]{aes}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Diagnose multiple tSNE runs with the correlations of coordinates.
}
