% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect-explore.R
\name{explore_batch_effect_meth_rle}
\alias{explore_batch_effect_meth_rle}
\title{Explore batch effect meth - RLE.}
\usage{
explore_batch_effect_meth_rle(
  trainer,
  batch_name = "Batch",
  top_n = 20000,
  n_samples = 100
)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}

\item{batch_name}{column name of batch in the \code{targets} attribute of
\code{YamatClassifierTrainer} object.}

\item{top_n}{an integer of the most variable N loci for relative log
expression (RLE) analysis.}

\item{n_samples}{if there are many samples, down-sample it to a number
while keep the proportion of each group.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Explore batch effect meth - RLE.
}
