% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect-explore-unmeth.R
\name{explore_batch_effect_unmeth_pca_compute}
\alias{explore_batch_effect_unmeth_pca_compute}
\title{Perform PCA on unmeth}
\usage{
explore_batch_effect_unmeth_pca_compute(
  trainer,
  batch_name,
  top_n = 20000,
  k = 50,
  seed = 1,
  threshold = 0.9,
  unmeth_pca_rda = "unmeth_pca.Rda"
)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}

\item{batch_name}{column name of batch in the \code{targets} attribute of
\code{YamatClassifierTrainer} object.}

\item{top_n}{an integer of the most variable N loci for relative log
expression (RLE) analysis.}

\item{k}{Number of eigenvalues requested.}

\item{seed}{A numeric number of seed used for Capper's unmethod of
determining PC number. In brief, the unmethod shuffles features
across samples and determine the PC number by comparing the
maximum of eigen values from the randomization.}

\item{threshold}{A numeric scalar between 0 to 1 of the threshold of
the fraction of variance to choose PC number. Default to 0.9.}

\item{unmeth_pca_rda}{Rda file for unmeth PCA result.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Perform PCA on unmeth
}
