% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{find_pc_number.capper}
\alias{find_pc_number.capper}
\title{Find PC number: Capper's method.}
\usage{
find_pc_number.capper(x, eigen_values)
}
\arguments{
\item{x}{A matrix which has columns as features and rows as samples.}

\item{eigen_values}{A vector of eigen values. It is the item
\code{values} of the item \code{eigs} of the returned value of
\code{\link{pca123}}, which is returned by
\code{\link[RSpectra]{eigs}}.}
}
\value{
A list of three items:
  \itemize{
    \item \code{pc_num} An integer scalar of PC number.
    \item \code{variance_fraction} The sum of fraction of variance
      of \code{pc_num} PCs.
   \item \code{eigs_shuffled} A numeric vector of eigen values of
      the randomized data.
    \item \code{plot} A \code{\link[ggplot2]{ggplot}} object of
      density plot of eigenvalues of observed and randomized data.
  }
}
\description{
The method is described in the paper by Capper et al. entitled
"DNA methylation-based classification of central nervous system
tumours". In brief, the method shuffles features across samples
and determine the PC number by comparing the maximum of eigen
values from the randomization.
}
