% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{find_pc_number.var_frac}
\alias{find_pc_number.var_frac}
\title{Find PC number: fraction of variance.}
\usage{
find_pc_number.var_frac(eigen_values, threshold = 0.9)
}
\arguments{
\item{eigen_values}{A vector of eigen values. It is the item
\code{values} of the item \code{eigs} of the returned value of
\code{\link{pca123}}, which is returned by
\code{\link[RSpectra]{eigs}}.}

\item{threshold}{A numeric scalar between 0 to 1 of the threshold of
the fraction of variance. Default to 0.9.}
}
\value{
Depends on which method is called.
}
\description{
The fraction of total variance retained for each PC is calculated by
dividing each eigen value by the total sum of eigen values. The
assumption behind the approach is that high variance surrogates
high importance, which may be not true in the real data.
}
