% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{find_pc_number.yamat}
\alias{find_pc_number.yamat}
\title{Find PC number: yamat method.}
\usage{
find_pc_number.yamat(x, eigen_values, n = 30)
}
\arguments{
\item{x}{A matrix which has columns as features and rows as samples.}

\item{eigen_values}{A vector of eigen values. It is the item
\code{values} of the item \code{eigs} of the returned value of
\code{\link{pca123}}, which is returned by
\code{\link[RSpectra]{eigs}}.}

\item{n}{An integer scalar of randomization times. Default to 30.}
}
\value{
A list of three items:
  \itemize{
    \item \code{pc_num} An integer scalar of PC number.
    \item \code{variance_fraction} The sum of fraction of variance
      of \code{pc_num} PCs.
    \item \code{raw} A \code{data.frame} of maximum eigen values
      and PC numbers.
  }
}
\description{
The method is based upon Capper's method (see \code{\link{find_pc_number.capper}}).
It estimates the number of the maximum eigenvalue of randomized data by
repeat the randomization for several times.
}
\note{
The Capper and yamat methods gives the same number of PCs
  on solid tumor example.
}
