% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_batch_effect_explore_dir}
\alias{get_batch_effect_explore_dir}
\title{Get the directory of batch effect exploratory analysis.}
\usage{
get_batch_effect_explore_dir(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
path of the directory of batch effect exploratory analysis.
}
\description{
Get the directory of batch effect exploratory analysis.
}
