% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect-correction.R
\name{get_batch_matrix}
\alias{get_batch_matrix}
\title{Get the batch matrix for fitting batch effect correction.}
\usage{
get_batch_matrix(batch = NULL, batch2 = NULL, covariates = NULL)
}
\arguments{
\item{batch}{factor or vector indicating batches.}

\item{batch2}{factor or vector indicating a second series of batches.}

\item{covariates}{matrix or vector of numeric covariates to be adjusted for.}
}
\value{
a matrix for batch.
}
\description{
Get the batch matrix for fitting batch effect correction.
}
\details{
Inspired by \code{\link[limma]{removeBatchEffect}}.
}
