% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_beta}
\alias{get_beta}
\title{Get beta value.}
\usage{
get_beta(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
a matrix of beta value.
}
\description{
Get beta value.
}
