% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_beta_value_adjusted}
\alias{get_beta_value_adjusted}
\title{Get adjusted beta value.}
\usage{
get_beta_value_adjusted(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
a matrix of beta value.
}
\description{
Get adjusted beta value.
}
