% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_beta_value_adjusted_rda}
\alias{get_beta_value_adjusted_rda}
\title{Get adjusted beta value Rda file path.}
\usage{
get_beta_value_adjusted_rda(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
path of adjusted beta value Rda file.
}
\description{
Get adjusted beta value Rda file path.
}
