% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-plot.R
\name{get_main_pca_plot}
\alias{get_main_pca_plot}
\title{Main PCA plot}
\usage{
get_main_pca_plot(
  data,
  pal,
  batch_label = "Batch",
  classification_label = "Classification",
  x_label = "PC1",
  y_label = "PC2",
  legend_cex = 0.7
)
}
\description{
Main PCA plot
}
