% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_meth}
\alias{get_meth}
\title{Get meth.}
\usage{
get_meth(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
a matrix of meth
}
\description{
Get meth.
}
