% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_preprocessed_dir}
\alias{get_preprocessed_dir}
\title{Get preprocessed file directory.}
\usage{
get_preprocessed_dir(trainer, dir_name = "dkfz_preprocessed")
}
\description{
Get preprocessed file directory.
}
