% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_probe_ids_rda}
\alias{get_probe_ids_rda}
\title{Get probe ID file directory.}
\usage{
get_probe_ids_rda(trainer, file_name = "probe_ids.Rda")
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}

\item{file_name}{output file name.}
}
\value{
path of probe ID file.
}
\description{
Get probe ID file directory.
}
