% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R
\name{get_probes}
\alias{get_probes}
\title{Update probes of interest.}
\usage{
get_probes(trainer, chip_type_name = NULL, present_by_epic_v2 = TRUE)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}

\item{chip_type_name}{column contains the chip type name.}

\item{present_by_epic_v2}{bool if only include probes present in EPIC v2.}
}
\value{
a list of probe CG IDs.
}
\description{
If probe set is not set for \code{YamatClassifierTrainer}, check if a Rda
file exists for probe set. If exists, load Rda file. Otherwise, get probe IDs
from scratch. The probe set has:
  \itemize{
    \item Drop SNP.
    \item Drop loci on chr X or Y..
    \item Common between the array types of trainer and EPIC v2
  }
}
\details{
If probe set is set, returns it and meanwhile save it in Rda file if the
file does not exist.
}
