% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R
\name{get_probes_from_scratch}
\alias{get_probes_from_scratch}
\title{Get probes of interest}
\usage{
get_probes_from_scratch(trainer, chip_type_name, present_by_epic_v2 = TRUE)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}

\item{chip_type_name}{column contains the chip type name.}

\item{present_by_epic_v2}{bool if only include probes present in EPIC v2.}
}
\value{
a list of probe CG IDs.
}
\description{
The probe set has:
  \itemize{
    \item Drop SNP.
    \item Drop loci on chr X or Y..
    \item Common between the array types of trainer and EPIC v2
  }
}
