% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_targets}
\alias{get_targets}
\title{Get targets.}
\usage{
get_targets(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
\code{data.frame} of targets. See \code{\link[minfi]{read.metharray.exp}}.
}
\description{
Get targets.
}
