% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_trainer_rda}
\alias{get_trainer_rda}
\title{Get YamatClassifierTrainer Rda file.}
\usage{
get_trainer_rda(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
path of YamatClassifierTrainer Rda file.
}
\description{
Get YamatClassifierTrainer Rda file.
}
