% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_unmeth}
\alias{get_unmeth}
\title{Get unmeth.}
\usage{
get_unmeth(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
a matrix of unmeth
}
\description{
Get unmeth.
}
