% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_unmeth_from_preprocessed_files}
\alias{get_unmeth_from_preprocessed_files}
\title{Get unmeth from preprocessed files.}
\usage{
get_unmeth_from_preprocessed_files(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
a matrix of unmeth
}
\description{
Get unmeth from preprocessed files.
}
