% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{multiplot_tsne}
\alias{multiplot_tsne}
\title{Plot multiple tSNE results.}
\usage{
multiplot_tsne(x, pheno, ...)
}
\arguments{
\item{x}{A list of the returned value of \code{\link[RTsne]{RTsne}}
which itself is a list.}

\item{pheno}{A character or factor vector.}

\item{...}{Any arguments passed to \code{\link[ggplot2]{aes}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot multiple tSNE results.
}
