% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-plot.R
\name{plot_pca_result}
\alias{plot_pca_result}
\title{Plot PCA.}
\usage{
plot_pca_result(
  pca_result,
  pheno,
  batch_name,
  classification_name,
  pc_x = 1,
  pc_y = 2,
  batch_label = "Batch",
  classification_label = "Classification",
  density_lwd = 0.2
)
}
\arguments{
\item{pca_result}{An object of pca_result class.}

\item{pheno}{the \code{data.frame} of phenotype.}

\item{batch_name}{the column name of batch in the phenotype \code{data.frame}.}

\item{classification_name}{the column name of classification in the phenotype
\code{data.frame}.}

\item{threshold}{A numeric scalar between 0 to 1 of the threshold of
the fraction of variance to choose PC number. Default to 0.9.}
}
\value{
A list of four elements:
  \itemize{
    \item \code{projected} The result matrix of PCA analysis.
    \item \code{pca123} A list of result from step 1-3 returned by
      \code{\link{pca123}}.
    \item \code{capper} A list of the result of choosing PC number
      by Capper's method, returned by \code{\link{find_pc_number.capper}}.
    \item \code{vf} A list of the result of choosing PC number
      by fraction of variance, returned by \code{\link{find_pc_number.var_frac}}.
  }
}
\description{
Plot PCA.
}
\details{
The function wraps up the following five steps of PCA:
   \enumerate{
    \item Center and scale.
    \item Compute the correlation/covariance matrix.
    \item Calculate the eigenvectors and eigenvalues.
    \item Choose the PC number. I use Capper's method and fraction of
      variance to calculate PC numbers and choose the bigger one
      from the two methods. See details at \code{\link{find_pc_number.capper}}
      and \code{\link{find_pc_number.var_frac}}.
    \item Project the scaled input matrix onto the new basis.
  }
  I use \code{\link[RSpectra]{eigs}} function in Rspectra package
  instead of \code{\link[base]{eigen}} function in base to deal
  with large matrix.
}
