% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative-log-expression.R
\name{plot_relative_log_expression}
\alias{plot_relative_log_expression}
\title{Plot relative log expression.}
\usage{
plot_relative_log_expression(
  x,
  pheno,
  batch_name,
  batch_label,
  plot_title,
  plot_subtitle,
  seed = 123,
  n_samples = 100
)
}
\arguments{
\item{x}{A matrix of RLE (RLE).}

\item{batch_name}{column name of batch in the \code{targets} attribute of
\code{YamatClassifierTrainer} object.}

\item{batch_label}{batch label in the ggplot (color).}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle}

\item{seed}{seed for sampling.}

\item{n_samples}{if there are many samples, down-sample it to a number
while keep the proportion of each group.}

\item{top_n}{an integer of the most variable N loci for relative log
expression (RLE) analysis.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot relative log expression.
}
