% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot tSNE result.}
\usage{
plot_tsne(tsne_res, ...)
}
\arguments{
\item{tsne_res}{The returned value of \code{\link[RTsne]{RTsne}},
which is a list.}

\item{...}{Any arguments passed to \code{\link[ggplot2]{aes}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot tSNE result.
}
