% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dkfz.R
\name{preprocess_dkfz}
\alias{preprocess_dkfz}
\title{Preprocess DKFZ approach.}
\usage{
preprocess_dkfz(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
output Rda files of preprocessed data of \code{\link[minfi]{MethySet}}
  class. One file per chip.
}
\description{
Preprocess DKFZ approach.
}
