% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{project_pc}
\alias{project_pc}
\title{Project to new basis.}
\usage{
project_pc(pca123_res, pc_num)
}
\arguments{
\item{pca123_res}{A list returned by \code{\link{pca123}}.}

\item{pc_num}{An integer scalar of PC number.}
}
\value{
A matrix which has columns as new features (PCs) and rows
  as samples.
}
\description{
Project the scaled input matrix (i.e. z-scores) onto the new basis.
}
