% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection.R
\name{select_features_ranger}
\alias{select_features_ranger}
\title{Perform random forest feature selection with ranger package.}
\usage{
select_features_ranger(
  dat,
  response_name,
  top_n = 10000,
  num_trees = 100,
  importance = "permutation",
  ...
)
}
\arguments{
\item{dat}{a \code{data.frame} of input data.}

\item{response_name}{column name of the response.}

\item{top_n}{most important N features.}

\item{num_trees}{Number of trees. Default to 100 instead of 500 in \code{\link[ranger]{ranger}}.}

\item{importance}{see \code{\link[ranger]{ranger}} importance argument. Default to "permutation" here.}

\item{...}{parameters for \code{\link[ranger]{ranger}}.}
}
\value{
a vector of selected features.
}
\description{
Perform random forest feature selection with ranger package.
}
