% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{shuffle_matrix}
\alias{shuffle_matrix}
\title{Shuffle matrix.}
\usage{
shuffle_matrix(x, margin = c("column", "row"))
}
\arguments{
\item{x}{A matrix.}

\item{margin}{A character scalar, either "row" or "column". Default
to "column", because columns are usually used as features.}
}
\value{
A matrix.
}
\description{
Each row or column is shuffled separately. To reproduce the result,
set the seed before running the function.
}
