% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-effect-correction.R
\name{train_batch_effect_model}
\alias{train_batch_effect_model}
\title{Train the model for batch effect removal.}
\usage{
train_batch_effect_model(
  x,
  batch = NULL,
  batch2 = NULL,
  covariates = NULL,
  design = matrix(1, ncol(x), 1),
  log2_transform = TRUE,
  adjusted_rda = "adjusted.Rda",
  fit_rda = "fit.Rda",
  ...
)
}
\arguments{
\item{x}{a matrix of log2 transformed signal intensity.}

\item{batch}{factor or vector indicating batches.}

\item{batch2}{factor or vector indicating a second series of batches.}

\item{covariates}{matrix or vector of numeric covariates to be adjusted for.}

\item{design}{design matrix relating to treatment conditions to be
preserved, usually the design matrix with all experimental factors other
than the batch effects.}

\item{log2_transform}{if perform log2 transform with offset 1, i.e. formula
log2(x+1).}

\item{adjusted_rda}{the file save the adjusted value.}

\item{fit_rda}{the file save the model.}

\item{...}{parameters pass to \code{\link[limma]{lmFit}}.}
}
\value{
a list of fit result.
}
\description{
Train the model for batch effect removal.
}
\details{
Inspired by \code{\link[limma]{removeBatchEffect}}.
}
