% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model5.R
\name{train_calibration_model_ridge}
\alias{train_calibration_model_ridge}
\title{Train calibration model with ridge regularization.}
\usage{
train_calibration_model_ridge(
  X,
  y,
  youden_index_threshold = 0.9,
  lambda_min_ratio = 1e-06
)
}
\arguments{
\item{X}{a matrix of predicted values.}

\item{y}{a factor vector of response.}

\item{youden_index_threshold}{the threshold. Default to 0.9.}

\item{lambda_min_ratio}{see \code{lambda.min.ratio} of \code{\link[glmnet]{glmnet}}.}
}
\value{
calibration result.
}
\description{
Train calibration model with ridge regularization.
}
