% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model4.R
\name{train_model4_outer_fold}
\alias{train_model4_outer_fold}
\title{Train model 4 - Process the outer fold}
\usage{
train_model4_outer_fold(
  dat,
  response_name,
  outer_train_index,
  random_state,
  inner_cv_folds = 5,
  rf_grid = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{a \code{data.frame} of input data.}

\item{response_name}{column name of the response.}

\item{outer_train_index}{outer train sample indexes.}

\item{random_state}{random seed.}

\item{inner_cv_folds}{inner cross-validation fold number.}

\item{rf_grid}{A data frame with possible tuning values. See \code{\link[caret]{train}}.}

\item{verbose}{A bool.}
}
\value{
a list of two attributes, a \code{data.frame} of calibrated
  probabilities and response variable, and a vector of selected features.
}
\description{
Train model 4 - Process the outer fold
}
