% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model5.R
\name{train_model5_inner_fold}
\alias{train_model5_inner_fold}
\title{Train model 5 - process the inner fold}
\usage{
train_model5_inner_fold(
  outer_train,
  response_name,
  random_state,
  inner_cv_folds = 5,
  calibration_youden_index_threshold = 0.9,
  calibration_lambda_min_ratio = 1e-06,
  rf_grid = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{outer_train}{a \code{data.frame} of one fold of training set during
nested cross-validation.}

\item{response_name}{column name of the response.}

\item{random_state}{random seed.}

\item{inner_cv_folds}{inner cross-validation fold number.}

\item{calibration_youden_index_threshold}{a float of Youden index threhold.
Default to 0.9.}

\item{calibration_lambda_min_ratio}{see \code{lambda.min.ratio} of
\code{\link[glmnet]{glmnet}}. Default to 1e-6.}

\item{rf_grid}{A data frame with possible tuning values. See \code{\link[caret]{train}}.}

\item{verbose}{A bool.}
}
\value{
a list of two models of random forest and calibration model respectively.
}
\description{
Train model 5 - process the inner fold
}
