% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model5.R
\name{train_model5_outer_fold}
\alias{train_model5_outer_fold}
\title{Train model 5 - Process the outer fold}
\usage{
train_model5_outer_fold(
  dat,
  response_name,
  outer_train_index,
  random_state,
  inner_cv_folds = 3,
  calibration_youden_index_threshold = 0.9,
  calibration_lambda_min_ratio = 1e-06,
  rf_grid = NULL,
  result_file = "outer_fold.Rda",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{a \code{data.frame} of input data.}

\item{response_name}{column name of the response.}

\item{outer_train_index}{outer train sample indexes.}

\item{random_state}{random seed.}

\item{inner_cv_folds}{inner cross-validation fold number.}

\item{calibration_youden_index_threshold}{a float of Youden index threhold.
Default to 0.9.}

\item{calibration_lambda_min_ratio}{see \code{lambda.min.ratio} of
\code{\link[glmnet]{glmnet}}. Default to 1e-6.}

\item{rf_grid}{A data frame with possible tuning values. See \code{\link[caret]{train}}.}

\item{result_file}{save result in Rda file.}

\item{verbose}{A bool.}
}
\value{
a list of two attributes, a \code{data.frame} of calibrated
  probabilities and response variable, and a vector of selected features.
}
\description{
Train model 5 - Process the outer fold
}
