% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_preprocess_pipe.R
\name{train_preprocess_pipe}
\alias{train_preprocess_pipe}
\title{Train preprocess pipeline}
\usage{
train_preprocess_pipe(idat_dir, targets, output, overwrite = FALSE)
}
\arguments{
\item{idat_dir}{IDAT file directory.}

\item{targets}{\code{data.frame} of targets. See \code{\link[minfi]{read.metharray.exp}}.}

\item{output}{output directory.}

\item{overwrite}{A bool if overwrite the result files.}
}
\value{
A S3 object of \code{YamatClassifierTrainer} class.
}
\description{
Preprocess for the training/reference data set from IDAT file to beta values.
Notice batch effect correction is not included.
}
