% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_rf_ridge}
\alias{train_rf_ridge}
\title{Train and evaluate the model with random forest and ridge-penalized
multinomial logistic regression (MR)}
\usage{
train_rf_ridge(
  dat,
  response_name,
  feature_selection = c("Boruta"),
  outer_cv_folds = 5,
  inner_cv_folds = 5,
  random_state = 56,
  rf_grid = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{a \code{data.frame} of input data.}

\item{response_name}{column name of the response.}

\item{feature_selection}{feature selection method.}

\item{outer_cv_folds}{outer cross-validation fold number.}

\item{inner_cv_folds}{inner cross-validation fold number.}

\item{random_state}{random seed.}

\item{rf_grid}{A data frame with possible tuning values. See \code{\link[caret]{train}}.}

\item{verbose}{A bool.}
}
\value{
To be added.
}
\description{
Train and evaluate the model with random forest and ridge-penalized
multinomial logistic regression (MR)
}
