% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_rf_ridge_outer_fold}
\alias{train_rf_ridge_outer_fold}
\title{Process the outer fold}
\usage{
train_rf_ridge_outer_fold(
  dat,
  response_name,
  outer_train_index,
  random_state,
  feature_selection = c("Boruta"),
  inner_cv_folds = 5,
  rf_grid = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{a \code{data.frame} of input data.}

\item{response_name}{column name of the response.}

\item{outer_train_index}{outer train sample indexes.}

\item{random_state}{random seed.}

\item{feature_selection}{feature selection method.}

\item{inner_cv_folds}{inner cross-validation fold number.}

\item{rf_grid}{A data frame with possible tuning values. See \code{\link[caret]{train}}.}

\item{verbose}{A bool.}
}
\value{
a \code{data.frame} of calibrated probabilities and response variable.
}
\description{
Process the outer fold
}
