% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tsne}
\alias{tsne}
\title{Run tSNE in batch mode.}
\usage{
tsne(x, n = 10, ...)
}
\arguments{
\item{x}{A matrix.}

\item{n}{An integer scalar. Default to 10.}

\item{...}{Any arguments passed to \code{\link[RTsne]{RTsne}}.}
}
\value{
A list of the returned value of \code{\link[RTsne]{RTsne}}.
}
\description{
Run tSNE in batch mode.
}
