% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{create_trainer}
\alias{create_trainer}
\title{Create a YamatClassifierTrainer object.}
\usage{
create_trainer(idat_dir, targets, output, probes = NULL, overwrite = FALSE)
}
\arguments{
\item{idat_dir}{IDAT file directory.}

\item{targets}{\code{data.frame} of targets. See \code{\link[minfi]{read.metharray.exp}}.}

\item{output}{output directory.}

\item{probes}{a character vector of probes.}

\item{overwrite}{A bool if overwrite the result files.}
}
\value{
A S3 object of \code{YamatClassifierTrainer} class.
}
\description{
Create a YamatClassifierTrainer object.
}
