% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trainer.R, R/train.R
\name{create_trainer}
\alias{create_trainer}
\title{Create a YamatClassifierTrainer object.}
\usage{
create_trainer(idat_dir, targets, output, overwrite = FALSE)

create_trainer(idat_dir, targets, output, overwrite = FALSE)
}
\arguments{
\item{idat_dir}{IDAT file directory.}

\item{targets}{\code{data.frame} of targets. See \code{\link[minfi]{read.metharray.exp}}.}

\item{output}{output directory.}

\item{overwrite}{A bool if overwrite the result files.}
}
\value{
A S3 object of YamatClassifierTrainer class.

A S3 object of YamatClassifierTrainer class.
}
\description{
Create a YamatClassifierTrainer object.

Create a YamatClassifierTrainer object.
}
