% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{get_preprocessed_dir}
\alias{get_preprocessed_dir}
\title{Get preprocessed file directory.}
\usage{
get_preprocessed_dir(trainer)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}
}
\value{
path of preprocessed file directory.
}
\description{
Get preprocessed file directory.
}
