% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-plot.R
\name{plot_pca_result}
\alias{plot_pca_result}
\title{Plot PCA.}
\usage{
plot_pca_result(
  pca_result,
  pheno,
  batch_name,
  classification_name,
  pc_x = 1,
  pc_y = 2,
  batch_label = "Batch",
  classification_label = "Classification",
  density_lwd = 0.6
)
}
\arguments{
\item{pca_result}{An object of pca_result class.}

\item{pheno}{the \code{data.frame} of phenotype.}

\item{batch_name}{the column name of batch in the phenotype \code{data.frame}.}

\item{classification_name}{the column name of classification in the phenotype
\code{data.frame}.}

\item{pc_x}{an integer of PC on x axis. Default to 1.}

\item{pc_y}{an integer of PC on y axis. Default to 2.}

\item{batch_label}{a character describe the ggplot2 label for batch.}

\item{classification_label}{a character describe the ggplot2 label for
classification.}
}
\value{
A \code{\link[gridExtra]{tableGrob}} object.
}
\description{
Plot PCA.
}
\details{
Inspired by PLSDAbatch package.
}
