% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-plot.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare PCA plot data}
\usage{
prepare_plot_data(
  projection,
  pheno,
  batch_name,
  classification_name,
  pc_x_name = "PC1",
  pc_y_name = "PC2"
)
}
\description{
Prepare PCA plot data
}
