% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{run_rtsne}
\alias{run_rtsne}
\title{Computing t-SNE with \code{\link[Rtsne]{Rtsne}}.}
\usage{
run_rtsne(
  trainer,
  top_n = 10000,
  perplexity = 30,
  n_iter = 3000,
  random_state = 123,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{trainer}{A S3 object of \code{YamatClassifierTrainer} class.}

\item{top_n}{an integer of the most variable N loci for t-SNE}

\item{perplexity}{Numeric scalar controlling the neighborhood used when
estimating the embedding. Default to 30.}

\item{n_iter}{Integer scalar specifying the number of iterations to complete.
Default to 3000.}

\item{random_state}{Integer scalar specifying the seed used by the random
number generator.}

\item{...}{other arguments of \code{\link[Rtsne]{Rtsne}}.}
}
\value{
A matrix of t-SNE embeddings.
}
\description{
Perform t-SNE with fit-SNE algorithm using \code{\link[Rtsne]{Rtsne}}.
}
\details{
Notice set \code{pca} argument as default (TRUE).
}
